src/EducativeIo/Courses/GrokkingTheCodingInterview/Ch14_TopKElements/PC3_FrequencyStack/Java/SolutionTest.java
package EducativeIo.Courses.GrokkingTheCodingInterview.Ch14_TopKElements.PC3_FrequencyStack.Java;

import static org.junit.jupiter.api.Assertions.*;
import java.time.Duration;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

public class SolutionTest {

    Solution solution;

    @BeforeEach
    public void setUp() throws Exception {
        solution = new Solution();
    }

    @AfterEach
    public void tearDown() throws Exception {
        solution = null;
    }

    @Test
    public void MainFunction() {
        assertTimeout(Duration.ofMillis(700), () -> {
            String[] args = new String[0];
            assertAll(() -> Solution.main(args));
        });
    }

    @Test
    public void TrivialCase1() {
        Solution frequencyStack = new Solution();
        frequencyStack.push(1);
        frequencyStack.push(2);
        frequencyStack.push(3);
        frequencyStack.push(2);
        frequencyStack.push(1);
        frequencyStack.push(2);
        frequencyStack.push(5);
        assertTimeout(Duration.ofMillis(700), () -> {
            assertEquals(2, frequencyStack.pop());
            assertEquals(1, frequencyStack.pop());
            assertEquals(2, frequencyStack.pop());
        });
    }
}
